Move lib to be top level
